Add src folder, change all imports and remove .vscode file
