+++ .gitignore
@@ -34,3 +34,6 @@ yarn-error.log*
 
 # vercel
 .vercel
+
+# vscode
+.vscode/*
