--- .vscode/launch.json
@@ -1,15 +1,0 @@
-{
-  // Verwendet IntelliSense zum Ermitteln möglicher Attribute.
-  // Zeigen Sie auf vorhandene Attribute, um die zugehörigen Beschreibungen anzuzeigen.
-  // Weitere Informationen finden Sie unter https://go.microsoft.com/fwlink/?linkid=830387
-  "version": "0.2.0",
-  "configurations": [
-    {
-      "type": "pwa-chrome",
-      "request": "launch",
-      "name": "Launch Chrome against localhost",
-      "url": "http://localhost:8080",
-      "webRoot": "${workspaceFolder}"
-    }
-  ]
-}
