+++ src/components/Card.js
@@ -1,6 +1,8 @@
 import { React, useState } from 'react';
-import styles from 'styles/card.module.css';
+
 import AuthCard from 'components/AuthCard';
+
+import styles from 'styles/card.module.css';
 
 export default function Card(props) {
   const [enteringPhaseOne, setEnteringPhaseOne] = useState(false);
