+++ src/components/WalletStatus.js
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import styles from 'styles/walletstatus.module.css';
 
 export default function WalletStatus() {
