+++ src/components/icons/walletOutline.js
