+++ src/components/icons/walletPlusOutline.js
